@@ -1,9 +1,9 @@
 class SpudPostSite < ActiveRecord::Base
-  attr_accessible :spud_post_id, :spud_site_id
-  belongs_to :spud_post
+  # attr_accessible :spud_post_id, :spud_site_id
+  belongs_to :spud_post, :touch => :true
 
   def spud_site
     return Spud::Core.site_config_for_id(spud_site_id)
   end
-  
+
 end
